Refresh feed works
