+++ Gemfile
@@ -3,6 +3,7 @@ source 'https://rubygems.org'
 # PostgreSQL driver
 gem 'pg'
 gem 'pry'
+gem 'hirb'
 
 # Sinatra driver
 gem 'sinatra'
