+++ app/views/layout.erb
@@ -18,6 +18,7 @@
   <link rel="stylesheet" href="/css/application.css">
 
   <script src="//ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
+  <script src="//cdnjs.cloudflare.com/ajax/libs/mustache.js/2.2.0/mustache.min.js"></script>
   <script src="/js/application.js"></script>
   <link href='https://fonts.googleapis.com/css?family=Josefin+Slab:700italic,700,600italic,600,400italic,400,300italic,300,100italic,100' rel='stylesheet' type='text/css'>
